+++ client/src/server-data.js
@@ -1,2 +1,3 @@
 // This is data that is defined by the server see Func:getServerData
-export default window.serverData
+module.exports = window.serverData
+module.exports.default = window.serverData